refact(MainWindow): The preset buttons are generated dynamically in the constructor of the window feat(SystemService.cs): add 3 events of logical axis to actuate the axis refact(DistanceUnitHelper.cs): add abs/rel distance to represent distance in real world unit refact(MoveArgs.cs): now derived from EventArgs feat(AxisBase.cs): add new move-relatived functions to support distance in real world unit refact(AxisBase.cs): change accessibility of UnitHelper property to protected set and get feat(LogicalAxis.cs): change actuation logic
Added Home/Move/Stop events in order to actuate the axis using the LogicalAxis class instance directly. The system service class raises the event to actuate the axis.
Also, the commands for the events are added in order to realize the mvvm pattern.

refact(LuminosAxis.cs): change the way of setting Dps property in the UnitHelper property
refact(Axis.xaml): change the way of binding of the current position textbox
refact(Axis.xaml.cs): change the way of move the axis
feat(ClassDiagram1.cd): create the class diagram